+++ src/v2i-hub/PriorityPlugin/src/PriorityPluginMessaging.cpp
@@ -69,7 +69,7 @@ namespace PriorityPlugin {
             if (bool shouldSet = (setData != _lastSentServiceRequest);
                 shouldSet) {
                 PLOG(logDEBUG3) << "PRS SET prsServiceRequest to CO: " << tmx::byte_stream_encode(setData);
-                if (bool setOk = SnmpSet(targetClient, NTCIP1211_PRS_SERVICE_REQUEST_OID, setData);
+                if (bool setOk = SnmpSet(targetClient, tsc::mib::ntcip1211::NTCIP1211_PRS_SERVICE_REQUEST_OID, setData);
                     !setOk) {
                     PLOG(logERROR) << "PRS failed to SET prsServiceRequest to CO";
                     std::this_thread::sleep_for(std::chrono::milliseconds(_pollIntervalMs));
@@ -88,7 +88,7 @@ namespace PriorityPlugin {
             while (coBusy && _running && maxRetries > 0) {
                 --maxRetries;
                 std::vector<uint8_t> getData;
-                if (bool getOk = SnmpGet(targetClient, NTCIP1211_PRS_SERVICE_REQUEST_OID, getData);
+                if (bool getOk = SnmpGet(targetClient, tsc::mib::ntcip1211::NTCIP1211_PRS_SERVICE_REQUEST_OID, getData);
                     !getOk) {
                     PLOG(logERROR) << "PRS failed to GET prsServiceRequest from CO";
                     coError = true;
@@ -141,7 +141,7 @@ namespace PriorityPlugin {
                 // j) SET prsServiceRequest to the CO only if the post-prioritization
                 // table differs from what the CO already holds.
                 if (updatedSetData != _lastSentServiceRequest) {
-                    if (!SnmpSet(targetClient, NTCIP1211_PRS_SERVICE_REQUEST_OID, updatedSetData)) {
+                    if (!SnmpSet(targetClient, tsc::mib::ntcip1211::NTCIP1211_PRS_SERVICE_REQUEST_OID, updatedSetData)) {
                         PLOG(logERROR) << "PRS failed to SET prsServiceRequest to CO after prioritization";
                     } else {
                         _lastSentServiceRequest = updatedSetData;
